Revert "Prescription Data ONLY parser"
This reverts commit ba75c6a450da599c75379833d65642c6ee865d9c.
